Distributed Cloud: enhance glance to support distributed keystone
This patch enhanced glance to support distributed keystone in
Distributed Cloud. It naturally fixed the issue where VIM fails to get
image list on subcloud.

The solution is to enable "use_user_token = False" support in glance
authentication so that glance-api on subcloud can access
glance-registry on System Controller by username and password. This is
neccessary because the token from subcloud keystone is not valide on
System Controller keystone.

The patch also fixed glance tox failures introduced by the changes.

Note: this is a temporary solution to make glance image available on
subcloud in Distributed Cloud. The change will be revisited once story
2002842(https://storyboard.openstack.org/#!/story/2002842) is in place.

Closes-Bug: 1791833

Signed-off-by: Andy Ning <andy.ning@windriver.com>